backend/requirements.txt
asgiref==3.7.2
Django==4.2.7
djangorestframework==3.14.0
pytz==2023.3.post1
sqlparse==0.4.4
certifi==2023.11.17
charset-normalizer==3.3.2
distlib==0.3.7
docopt==0.6.2
fonttools==4.45.1
idna==3.4
kiwisolver==1.4.5
lxml==4.9.3
numpy==1.26.1
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
pipreqs==0.4.13
platformdirs==3.11.0
polars==0.19.15
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
threadpoolctl==3.2.0
tzdata==2023.3
urllib3==2.1.0
yarg==0.1.9
